refactor: more react-doctor fixes
