+++ components/project-card.tsx
@@ -12,7 +12,13 @@ interface ProjectCardProps {
 
 export default function ProjectCard({ id, title, description, imageSrc, onClick }: ProjectCardProps) {
   return (
-    <div className="flex flex-col cursor-pointer transition-transform duration-300 hover:scale-102" onClick={onClick}>
+    <div 
+      role="button"
+      tabIndex={0}
+      className="flex flex-col cursor-pointer transition-transform duration-300 hover:scale-102" 
+      onClick={onClick}
+      onKeyDown={(e) => e.key === 'Enter' && onClick()}
+    >
       <div className="relative aspect-[4/3] w-full overflow-hidden rounded-md bg-gray-300">
         <Image src={imageSrc || "/placeholder.svg"} alt={title} fill sizes="(max-width: 640px) 100vw, (max-width: 1024px) 50vw, 33vw" className="object-cover" />
       </div>
