+++ components/timeline-rail.tsx
@@ -1,25 +1,72 @@
 "use client";
 
-import React, { useEffect, useState } from "react";
+import React, { useEffect, useReducer } from "react";
 
 interface TimelineRailProps {
   sectionRefs: React.RefObject<HTMLElement>[];
 }
 
-export default function TimelineRail({ sectionRefs }: TimelineRailProps) {
-  const [activeIndex, setActiveIndex] = useState(0);
-  const [filledDotTop, setFilledDotTop] = useState(0);
-  const [railLeft, setRailLeft] = useState<number>(64); // default ~4rem
-  const [markerTops, setMarkerTops] = useState<number[]>([]);
+interface TimelineRailState {
+  activeIndex: number;
+  filledDotTop: number;
+  railLeft: number;
+  markerTops: number[];
+}
 
-  // Calculate positions of empty circles based on section title positions
+type TimelineRailAction =
+  | { type: "SET_ACTIVE_INDEX"; payload: number }
+  | { type: "SET_FILLED_DOT_TOP"; payload: number }
+  | { type: "SET_RAIL_LEFT"; payload: number }
+  | { type: "SET_MARKER_TOPS"; payload: number[] }
+  | { type: "UPDATE_LAYOUT"; payload: { railLeft: number; markerTops: number[] } }
+  | { type: "UPDATE_POSITIONS"; payload: { activeIndex: number; filledDotTop: number } };
+
+const initialState: TimelineRailState = {
+  activeIndex: 0,
+  filledDotTop: 0,
+  railLeft: 64,
+  markerTops: [],
+};
+
+function timelineRailReducer(
+  state: TimelineRailState,
+  action: TimelineRailAction
+): TimelineRailState {
+  switch (action.type) {
+    case "SET_ACTIVE_INDEX":
+      return { ...state, activeIndex: action.payload };
+    case "SET_FILLED_DOT_TOP":
+      return { ...state, filledDotTop: action.payload };
+    case "SET_RAIL_LEFT":
+      return { ...state, railLeft: action.payload };
+    case "SET_MARKER_TOPS":
+      return { ...state, markerTops: action.payload };
+    case "UPDATE_LAYOUT":
+      return {
+        ...state,
+        railLeft: action.payload.railLeft,
+        markerTops: action.payload.markerTops,
+      };
+    case "UPDATE_POSITIONS":
+      return {
+        ...state,
+        activeIndex: action.payload.activeIndex,
+        filledDotTop: action.payload.filledDotTop,
+      };
+    default:
+      return state;
+  }
+}
+
+export default function TimelineRail({ sectionRefs }: TimelineRailProps) {
+  const [state, dispatch] = useReducer(timelineRailReducer, initialState);
+
   const computeLayout = () => {
     const positions: number[] = [];
-    // Align rail with the left edge of the first section container
     const first = sectionRefs[0]?.current;
     if (first) {
       const left = first.getBoundingClientRect().left;
-      setRailLeft(left);
+      dispatch({ type: "SET_RAIL_LEFT", payload: left });
     }
 
     sectionRefs.forEach((ref) => {
@@ -28,17 +75,15 @@ export default function TimelineRail({ sectionRefs }: TimelineRailProps) {
       const titleEl = el.querySelector("h2");
       if (!titleEl) return;
       const titleRect = titleEl.getBoundingClientRect();
-      // Use viewport-relative top for fixed container
       positions.push(titleRect.top);
     });
-    setMarkerTops(positions);
+    dispatch({ type: "SET_MARKER_TOPS", payload: positions });
   };
 
   useEffect(() => {
     const update = () => {
       if (sectionRefs.length === 0) return;
 
-      // Determine which section is most visible in viewport
       const viewportTop = window.scrollY;
       const viewportBottom = viewportTop + window.innerHeight;
 
@@ -53,7 +98,6 @@ export default function TimelineRail({ sectionRefs }: TimelineRailProps) {
         const sectionTop = rect.top + window.scrollY;
         const sectionBottom = sectionTop + rect.height;
 
-        // Calculate how much of the section is visible
         const visibleTop = Math.max(viewportTop, sectionTop);
         const visibleBottom = Math.min(viewportBottom, sectionBottom);
         const visibleHeight = Math.max(0, visibleBottom - visibleTop);
@@ -65,20 +109,22 @@ export default function TimelineRail({ sectionRefs }: TimelineRailProps) {
         }
       });
 
-      setActiveIndex(closestIdx);
-
-      // Get the current section's title position (viewport-relative)
       const currentSection = sectionRefs[closestIdx]?.current;
       if (currentSection) {
         const titleEl = currentSection.querySelector("h2");
         if (titleEl) {
           const titleRect = titleEl.getBoundingClientRect();
-          setFilledDotTop(titleRect.top);
+          dispatch({
+            type: "UPDATE_POSITIONS",
+            payload: {
+              activeIndex: closestIdx,
+              filledDotTop: titleRect.top,
+            },
+          });
         }
       }
     };
 
-    // Initial layout + update
     setTimeout(() => {
       computeLayout();
       update();
@@ -97,8 +143,6 @@ export default function TimelineRail({ sectionRefs }: TimelineRailProps) {
     };
   }, [sectionRefs]);
 
-  // Marker positions are kept in state (viewport-relative)
-
   return (
     <div
       className="pointer-events-none fixed left-0 top-0 bottom-0 z-10"
@@ -107,22 +151,22 @@ export default function TimelineRail({ sectionRefs }: TimelineRailProps) {
       {/* Continuous vertical line */}
       <div
         className="absolute top-0 bottom-0 w-0.5 bg-gray-200"
-        style={{ left: railLeft }}
+        style={{ left: state.railLeft }}
       />
 
       {/* Empty circles at each section title */}
-      {markerTops.map((topPos) => (
+      {state.markerTops.map((topPos) => (
         <div
           key={topPos}
           className="absolute -translate-x-1/2 h-4 w-4 rounded-full border-2 border-[#F5A623] bg-white"
-          style={{ top: topPos - 8, left: railLeft }}
+          style={{ top: topPos - 8, left: state.railLeft }}
         />
       ))}
 
       {/* Moving filled circle */}
       <div
         className="absolute -translate-x-1/2 h-4 w-4 rounded-full bg-[#F5A623] transition-[top] duration-300 ease-out z-10"
-        style={{ top: filledDotTop - 8, left: railLeft }}
+        style={{ top: state.filledDotTop - 8, left: state.railLeft }}
       />
     </div>
   );
